Creating ElectionManager to manage node status - Start loop created to handle state transitions - Created helper functions for setting candidate state / cycle no / term no - Created ElectionManager struct to hold key information of the node
Created nodeServices.go file
- Declaration of all gRPC services will reside in this file for RAFT protocol

Created ChannelsHelper.go file
- To store channels used for communication between gRPC and other go routines

Redesigned message structure for raft.go
- Added new components and services
- Removed messages pertaining to log
